+++ src/titiler/application/titiler/application/__init__.py
@@ -1,3 +1,3 @@
 """titiler.application"""
 
-__version__ = "2.0.0b2"
+__version__ = "2.0.0"