+++ src/titiler/xarray/titiler/xarray/__init__.py
@@ -1,3 +1,3 @@
 """titiler.xarray"""
 
-__version__ = "2.0.0b2"
+__version__ = "2.0.0"